requirements.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
future==0.17.1
idna==2.8
isort==4.3.17
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
matplotlib==3.0.3
mccabe==0.6.1
numpy==1.16.3
oauthlib==3.0.1
pandas==0.24.2
pylint==2.3.1
pyparsing==2.4.0
python-dateutil==2.8.0
python-twitter==3.5
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.24.2
wincertstore==0.2
wrapt==1.11.1
